Redesigned default media icon system. Old system pulled default graphics files as media and stored a copy in place of an actual preview. This was really really dumb: it made it difficult to change the icon after media load, caused icons to be distorted to fit whatever size the preview was specified for and required a fill set of icons for each supported mimetype.
The new system includes four sets of general icons: video, audio, document and queued media. Rather than import them, the fact than only a generic icon is available is stored in the media metadata and a suitable url returned. This means that changing icons is simple: just change out the generic icon files and, if needed, change the icon config to reflect the sizes you've created. The most appropriate size is always selected and used.