Implemented 'RDF mode' for exporter
This is useful for exports where you want have records of multiple
entities with different mappings next to each other (not nested). The
best example is building an RDF graph of objects, related entities,
list items, places etc., hence the name of the feature (for lack of a
better one). The configuration for this is pulled in via config file
(since it is some kind of meta exporter config that can include an
arbitrary number of mappings/exporters).