Implemented feature request PROV-486
The GeoNames data type now has two new options. They allow users to
customize exactly which elements are taken from the GeoNames search
result and added to the attribute value as text.

This is helpful if you want to use GeoNames to enable 'pseudo'
hierarchical geographical searches in your database. For instance, you
could have a display as detailed as this:

NA > United States > New York > New York City, Central Park
[40.78333,-73.96667]

enabling you to search for "Unites States" and still getting hits on
objects related to Central Park.

At the moment the configuration is a simple list of elements and a
delimiter. I thought about doing something like the display template
syntax with <ifdef> and co but that would have gotten out of hand very
quickly because the code for metadata display templates is not reusable.