PROV-495: Okay, I give in.
Make metadata element labels optional for sub-elements. It happens way
too often that there are nested Containers without labels which always
lead to problems with the ConfigurationExporter. This way it's
theoretically possible to create a label-less Text element inside a
container but I guess we're going to have to live with that.