Updating measurements with 'comma' issue
While editing/updating object measurements with comma the decimal part
of the value is lost;
This is caused when caConvertLocaleSpecificFloat returns string value
instead of a floating number value, to function
LengthAttributeValue::parseValue. The caConvertLocaleSpecificFloat
function converts a string value into float only if
function_exists("NumberFormatter") returns true. In case of false, it
returns with a string value;

In this fix we check if "NumberFormatter" function
exists(function_exists("NumberFormatter")) one level higher. We call
caConvertLocaleSpecificFloat only if the check is positive,
otherwise we convert the string value into the corresponding float
value.