PROV-870 Fix currency issue
.. that would prevent proper parsing of values returned by
getDisplayValue() (i.e. if a form with existing values is saved again)
because of weird multibyte characters in the currency formats returned
by Zend. We now replace their currency symbol placeholder with our own
and trim white spaces via regex, because trim() doesn’t seem to get
whatever they inserted there between the decimal placeholder and the
currency symbol placeholder. This was only a problem in locales where
the currency is a suffix with a space between the decimal and the
suffix (10,00 EUR).