Add mtime= URL parameter to local JS and CSS.
For local files, use filemtime() to grab the modification time
of the file and append it as a URL parameter.  This causes browsers
to cache the file until it is modified.  These files can therefore
be configured to be cached forever by the browser, as any modifications
will be automatically retrieved as there will be a new URL whenever the
file is modified.