PROV-922 Added support for relationship types
Previously it would be incredibly hard or even impossible to find out
the relationship type for items in sub-exports (what we call context
switches). We now support a couple of special keys in the source
columns. The names are along the lines of relationship display
templates:

* relationship_type_id
* relationship_type_code
* relationship_typename

Those work in deeply nested mappings as well and always refer to the
‘parent’ item in the export, if you will (the item where the context
was last switched). If the context is unchanged or these keys are
somehow not applicable, for instance if you traverse hierarchical
(type-less) relationships, mappings with these keys are ignored.