PROV-940 Changed config format to arrays of regexes
This makes it easier to have a parse-able config file and preg_replace
takes arrays as parameters for both search and replace anyway.

The whole thing is still very, very fragile because preg_replace is
very particular. We’re going to have to do a lot more sanity checking
before we actually pass the config parameters so that we can report
something useful to the user in case something goes wrong.