PROV-922 Proper localization for the datepicker
Unfortunately the localization files are set up in a way where loading
the corresponding js file is equivalent to setting the default to that
locale. That’s why we had to break it out in one package per locale in
javascript.conf, which we then can load when the attribute is rendered.
It’s a little excessive but it works.

Also, Canadians apparently disagree on their date format. The
CollectiveAccess DateTime localization file  for en_CA says the month
comes first in date displays (mm/dd/yy), the file for
jquery-ui.datepicker says it’s the same as en_GB (dd/mm/yy). Well, I
guess no one complained about the CollectiveAccess localization yet, so
I changed the format in the settings for the datepicker localization.