PROV-997 Added skipIfExpression setting to exporter
It uses the ExpressionParser that was written for the importer (see
http://docs.collectiveaccess.org/wiki/Expressions). You can reference
the value of the current mapping as variable ^value. Next we’re going
to implement a system that allows you to define your own variables in
exporter mappings which can then be referenced in these expressions. At
the moment they’re pretty useless.