PROV-998 Added variables to exporter
This allows you to use all the available exporter features to assign
dynamic values to an identifier and then later reference it as variable
in expressions for skipIfExpression. This is the only use so far but
potentially we could also allow using those as source or in templates
as well. The notion of saving something and then using it in a
different context later can be very useful in data exports.