PROV-985 Two steps back, one step forward
It seems to be impossible to insert arbitrary field content, in
particular broken HTML/XML, into existing DOM trees without CDATA
sections. This is a good thing! But: As a hack for the problem at hand,
we now added an option that explicitly strips the CDATA tags before
exporting. Doing this is not recommended and the option will probably
be undocumented indefinitely. It just doesn’t make sense in normal
scenarios.