PROV-1032 Changed representation annotation property interface to include a way to get a display value for all properties.
This is very useful for export where you don’t necessarily know what
kind of annotation you have at hand and want a list of properties
anyway. This is extensible enough so that we could add display format
options that are passed down to the coders if we need to.

This is a new function in the ca_representation_annotations BaseModel
API but can also be accessed using a ‘virtual’ property called
‘display’ like so:

$t_ann->get('ca_representation_annotations.props.display')

This makes it usable in display templates and the exporter.