PROV-1087 Always show element descriptions as tooltips
This mandated a bigger refactoring of the way we do tooltips. The
problem was that element containers with
minimumAttributeBundlesToDisplay=0 didn't have their element tooltips
displayed once you added an element in the browser because they were
managed statically by TooltipManager. We now add the JS code in the
bundle HTML so that it's executed when you click the "add element"
button.

To make sure we don't have tooltip JS code in two places, I added a
caGetTooltipJS() display helper that is now also used by
TooltipManager. I did not refactor the "tooltip functions" part yet
because I wasn't sure what that's about. It's not used anywhere, at
least not in Providence.