NOJIRA Migrate from ZipFile to PharData
Zipfile produces corrupt files on CentOS with PHP5.4 and I'm not in the
mood to reinvent the wheel so I thought we might as well migrate to
something that works.

The phar extension is not really a new requirement, we've been using it
in the zip extraction code of BaseModell::_processMedia for ages. Now
it's a formal requirement with config check and everything.

I'm not sure but I think the extension is part of the core distribution
since PHP5.3. It's available in the default php 5.4 packages in Debian
and CentOS. No extra setup required.