PROV-1167 Always index type ids were applicable
This is to prevent that filtering a result set on a list of type ids
doesn't yield any results. If type-level access control is enabled,
each and every query is restricted to a number of type ids. In some
cases, especially when removing a relationship, the type id for related
records would "disappear" from the search index, meaning the record
couldn't be found anymore (if type-based AC is enabled, that is).