NOJIRA Fix ElasticSearch list item type id filtering issue
Since list item type ids are optional, we have to search for
empty/unset values too. That is a little bit different in
ElasticSearch. Also, we have to avoid indexing an empty string for
these field values and leave them out of the index entirely instead.
It's a little hacky but it works. At some point we should rewrite this
using the ElasticSearch DSL and their PHP API.