NOJIRA Ignore type restrictions for subject table
The problem here is that sometimes you want to get() a type_id for a
record the current user doesn't have access to, to figure out if he has
access (e.g. in caCanRead()). But, if we return an empty value because
the user doesn't have access, he or she gains access. Argh ..

Thankfully this code will be replaced very soon anyway.