PROV-1268 Add token-based authentication to web service API
This finally allows us to re-use sessions for the API. While this is
technically not REST it's common practice and allows us to not basic
auth for every request.

The idea is that you call the /service.php/auth/login endpoint via GET
(w/ basic auth) and it'll give you an authToken. Append this authToken
as GET parameter to any subsequent service API calls and you'll be
authenticated and find yourself in the same session.

Auth tokens don't last very long (which is kind of the point) so you'e
going to have to re-authenticate every once in a while, when you get a
401.