PROV-1294 Move session lifetime handling into our code
This way we don't rely on external caching providers to get this right
and have an easier time debugging. We'll see if that gets rid of the
login/logout problems for good.

The cache items that store the session variables are now set to never
expire and the session lifetime is instead part of the session vars.