PROV-1317 Add prefetching to getWithTemplate
The assumption is that if you run getWithTemplate for the current row
in a search result, you'll probably run it for the next bunch of rows
too. Since running caProcessTemplateForIDs for every single row is
slow, we now prefetch a set number of rows when we first encounter a
template call. This should speed things up a little bit in Pawtucket
land.