PROV-1343 Fix bugs, complete missing features (save one – see below) and add tests
The "sort" and "sortDirection" options are currently unsupported. The old parser allowed sorting in a limited context and most users do no sorting at all. I'm looking into ways to expand this, or at least to implement it cleanly, perhaps as an option to SearchResult::get(), which would make it available in other places as well.