NOJIRA Template placeholder parser (caGetTemplateTags()) was pulling placeholders ending with ".%" and some imports use this in formatWithTemplate
Ex. ^/Last_Name, ^First_Name.%

would result in two placeholders for replacement: /LastName and /First_Name.%
The parser is smart enough to remove trailing periods and slashes but not %. % is a valid placeholder character, but only for separation of placeholder options from the placeholder itself. We now remove trailing % characters.