NOJIRA : allows to hook TimeExpressionParser preprocess
This allows me to develop into a plugin a preparser for French Revolutionary Calendar dates.
With the plugin activated, it would allow me to parse dates like "12 thermidor an II" (https://en.wikipedia.org/wiki/French_Republican_Calendar) before sending a normal date to the TEP class.

In the future, it would allow the development of plugins for dates in Hebrew, Islamic, persian, chinese calendars.