PROV-1405 Add recursive option to linked data config
Allows pulling literals from related nodes recursively, for instance
traversing broader relationships.

Note that this should be used with caution because we actually have to
load all these resources. Even though the results are cached for a
while, this can take a while when traversing large hierarchies and when
the Getty service happens to be slow.