PROV-1425 Turns out the bundle delete and interstitial editor code operates on relation_ids and we broke that with the last commit
This fix is a little ugly, because we have to translate that list of
relation_ids to a list of object_ids to make a search result. In
practice these lists shouldn't be longer than a few hundred so I think
this should work. If not, we can always craft a fast SQL query to get
the list of object_ids.

Oh yeah, we also revert ca_objects_results_list_html.php to the default
in this commit and add a second view with our custom mods for this
bundle. That list view now also always includes the relationship type.