PROV-1566 Yank old MemoryCache based setting caching
There's no reason for these dynamic methods to communicate via side
effects in some static container some place else. I get that we
potentially instantiate these a lot but pulling the settings from cache
once - when a new object is made - should be enough. This was just a
terrible rewrite of the caching code that predates MemoryCache&Co