PROV-1626 Use recommended way to hook into "beforeunload" event for unsaved changes warning
This doesn't fix the problem though. The event fires just fine in
Safari 9.1, but it refuses to show a dialog box.

It seems there was a recent change that could affect this:
https://developer.apple.com/library/mac/releasenotes/General/WhatsNewInS
afari/Articles/Safari_9_1.html?imm_mid=0deddd&cmp=em-web-na-na-newsltr_2
0160113

As I understand it, they want to replicate the Firefox behavior, which
means they display a generic message instead of a string set by the
script -- which would be fine, if it worked!