PROV-1619 Got started on sorting, but then found another issue with my relation_id -> related_id mapping code
I now added an option to BaseFindEngine::sort() that'll let us preserve
the indexes for the $pa_hits we pass in and then have
caMakeSearchResult() (optionally) return an index in addition to the
actual result.

This two-return-value thing is not pretty and I should probably build
that into the SearchResult itself. Maybe that's next.