Fix search parsing infinite loop / memory-overflow error.
When attempting to determine whether a given search includes Sets, it
is parsing the search but when there are parentheses `()` present, it
runs into an infinite loop.

This resolves the infinite loop by removing the parentheses from the
search text.  This doesn't actually affect whether or not the search
is for sets.