PROV-2376 Implement support for multiple mappings with a splitter attributes block. You can now do this:
{
	"placeType": "location",
	"relationshipType": "found",
	"attributes": {
		"idno": "^67",
		"plss": [{
			"township": "^70",
			"range": "^71",
			"section": "^72",
			"quarter": "^73"
		}, {
			"township": "^74",
			"range": "^75",
			"section": "^76",
			"quarter": "^77"
		}, {
			"township": "^78",
			"range": "^79",
			"section": "^80",
			"quarter": "^81"
		} ]
	}
}

Eg. the attributes "plss" container mapping, which previously could only be a single hash of sub-field values, can now be a list of hashes. This allows one to map from several different parts of an import data set into the same attributPROV-2376 Implement support for multiple mappings with a splitter attributes block. You can now do this:

{
	"placeType": "location",
	"relationshipType": "found",
	"attributes": {
		"idno": "^67",
		"plss": [{
			"township": "^70",
			"range": "^71",
			"section": "^72",
			"quarter": "^73"
		}, {
			"township": "^74",
			"range": "^75",
			"section": "^76",
			"quarter": "^77"
		}, {
			"township": "^78",
			"range": "^79",
			"section": "^80",
			"quarter": "^81"
		} ]
	}
}

Eg. the attributes "plss" container mapping, which previously could only be a single hash of sub-field values, can now be a list of hashes. This allows one to map from several different parts of an import data set into the same attribute