PROV-2988 In earlier days the search indexer supported INDEX_IDNO_PARTS and IDNO_DELIMITERS options which indexed identifier in parts based upon specified delimiters. For example, when INDEX_IDNO_PARTS was set and IDNO_DELIMITERS set to a list including ".",  2010.002.35 would be indexed as the discrete values 2010, 002 and 35.
These options were dropped when INDEX_AS_IDNO was added (although references to the old options were left in search_indexing.conf, causing confusion). INDEX_AS_IDNO uses multipart_id_numbering.conf configuration to index identifiers in a manner consistent with their configured format.

INDEX_AS_IDNO is superior in that it requires less configuration and indexes progessively, which is usually preferred. Thus 2010.002.35 is indexed as 2010.002.35, 2010.002 and 2010.

There are situation where forcing indexing on identifiers using split values is desirable. For example, if identifiers are configured as free text, but are typically set to delimited values, or in situations where you simply want to force all components of an identifer to be indexed IDNO_DELIMITERS can be handy.

IDNO_DELIMITERS should be restored as a sub-option of INDEX_AS_IDNO. When INDEX_AS_IDNO is set and  IDNO_DELIMITERS is set to a list of delimiters, identifier values will be split by those delimiters and the resulting values indexed separately.