Make ListSetByUser always show all set types
Another cheap fix proposal to un-confuse users listing sets by user... Make sure the menu selection for ListSetByUser always results in all set types being shown instead of depending on a context that may have captured choices of set type for previous listings by type. In general users are baffled when the same menu choice yields different listings. Or maybe this is a feature, and we're missing the point?