Update sets.css
Removed the move cursor from the .setItems a element and put it into the top .setItems class.  This is to make the cursor change when on the item and not if you add a <l> </l> into the sets item display template (which looks strange when the move icon only shows on the link back to the record).