Improve error checking for Shib field map
Previously error checking was almost non existant and performed its task
completely silently.
This change brings missing elements in the map right to the fore by
refusing to continue while uid and email are missing.

Rounding out the change is adding a stacktrace to "Could not create
SimpleSAML auth object", we've been carrying this change locally for
some time and its a great help in debugging.