Fix HTMLPurifier path in fix-permissions
PR 1587 introduced a change to allow configuring the HTML purifier
Serializer cache directory - that change omitted to update the
`fix-permissions` utility which has continued to try and update the old
path.
This commit updates `fix-permissions` to pull the cache directory from
the configuration file as other parts of CA do now.