Revert overly broad conversion of to https URLs
This reverts part of Seth's commit 2554940a77f2f3f6390b5b2c9e4486029b56acf8

The problem that commit was trying to fix is that for the Libary of
Congress, the request URL must always use https now.

However the URI which identifies the type of lookup (subject headings,
name authority, etc) must still use http, since that's part of the controlled identifier.

This change restores the original URIs which were correct. Because the
URI is used as the value of the vocabulary drop-down, this means that:
 (a) Users with pre-existing LCSH fields might actually work, because
they have the old value stored
 (b) Users who created a new LCSH field after the bad commit will
probably need to re-save the metadata element to set the correct URI

I don't know if this is a complete fix because there are other places in
the code where for example the URI is parsed out of the vocabulary
value, and those might need to be updated?

But at least for me, without this change, the autocomplete for LCSH
searches that specified a vocabulary would always fail with the latest
master code. With this change (and after re-saving the LCHS metadata
element to pick up the new value), autocomplete searches for LCSH vocabularies
seem to work